requirements.txt
Flask>=0.10.1
leancloud-sdk>=1.0.9
flask_bootstrap>=3.3.5.7